Fix the exponent bound being checked after it is narrowed
Reported as item 6 of #1059.

The exponent is accumulated into an int16_t and the bound is tested on
the result, so the wrap happens first and a wrapped value is under
MAX_EXP and is accepted. The digits after the wrap then decide what is
left, which makes the value the document gets a choice about.

    Oj::Parser.usual.parse("[1e327683]")   => [1000.0]
    Oj::Parser.usual.parse("[1e32768]")    => [1.0]
    Oj::Parser.usual.parse("[1e-327683]")  => [0.001]
    Oj::Parser.usual.parse("[1e-32768]")   => [1.0]

1e327683 comes back as 1000.0 because the accumulator wraps at the fifth
digit and the sixth, a 3, is all that survives.

The accumulator is an int now and only the value that passed the bound
is stored back, so the exponent either fits or the number goes to
big_change() the way one over the bound already did. Those four
documents now parse as the BigDecimals they are, and every exponent that
was already accepted is unchanged: 1e4932 is still Infinity, 1e4933 is
still 0.1e4934, and 1.5e10, 1e-5 and 2e-308 are untouched.

+++ ext/oj/parser.c
@@ -892,10 +892,10 @@ static void parse(ojParser p, const byte *json, size_t len, bool more) {
         case EXP_DIGIT:
             p->map = exp_map;
             for (; NUM_DIGIT == digit_map[*b]; b++) {
-                int16_t x = p->num.exp * 10 + (int16_t)(*b - '0');
+                int x = p->num.exp * 10 + (*b - '0');
 
                 if (x <= MAX_EXP) {
-                    p->num.exp = x;
+                    p->num.exp = (int16_t)x;
                 } else {
                     big_change(p);
                     p->map = big_exp_map;
